Rename phoneme page and fix render bugs
